Improve resilience to party changes during loading.
Make a shallow copy of the party list when loading starts, and use
that instead of the complete party list when waiting for a game to
start. This ensures we don't wait on new party members who are not
part of the current game. We also now re-evaluate everyone's load
state if we lose a member, in case they were the only one who hadn't
loaded yet.